Add TODOs, fix citation in `index.rst`, move `implicit_diff`
