+++ docs/initializers/linear.rst
@@ -3,6 +3,8 @@ ott.initializers.linear package
 .. currentmodule:: ott.initializers.linear
 .. automodule:: ott.initializers.linear
 
+.. TODO(marcocuturi): maybe add some text here
+
 Sinkhorn Initializers
 ---------------------
 .. autosummary::
