+++ docs/initializers/nn.rst
@@ -3,6 +3,8 @@ ott.initializers.nn package
 .. currentmodule:: ott.initializers.nn
 .. automodule:: ott.initializers.nn
 
+.. TODO(marcocuturi): maybe add some text here
+
 Neural Initializers
 -------------------
 .. autosummary::
