+++ ott/math/__init__.py
@@ -1,7 +1,6 @@
 from . import (
     decomposition,
     fixed_point_loop,
-    implicit_differentiation,
     matrix_square_root,
     unbalanced_functions,
     utils,
