+++ ott/solvers/linear/__init__.py
@@ -2,6 +2,7 @@ from . import (
     acceleration,
     continuous_barycenter,
     discrete_barycenter,
+    implicit_differentiation,
     sinkhorn,
     sinkhorn_lr,
 )
