+++ ott/solvers/linear/implicit_differentiation.py
