+++ ott/solvers/linear/sinkhorn.py
@@ -23,11 +23,10 @@ from typing_extensions import Literal
 
 from ott.geometry import geometry
 from ott.initializers.linear import initializers as init_lib
-from ott.math import fixed_point_loop
-from ott.math import implicit_differentiation as implicit_lib
-from ott.math import unbalanced_functions
+from ott.math import fixed_point_loop, unbalanced_functions
 from ott.problems.linear import linear_problem, potentials
 from ott.solvers.linear import acceleration
+from ott.solvers.linear import implicit_differentiation as implicit_lib
 
 __all__ = ["Sinkhorn", "SinkhornOutput"]
